refactor: extract android lifecycle helpers
